Moved listen path matcher to the router
